Create README.md
I analyzed the Twitter archive of We Rate Dogs through August 1, 2017. 
The standard Twitter data was augmented with data generated by image prediction software, 
as well as data about the stage of the dog's life (dog, puppy, or ect) generated by automated text extraction and organization programming. 

The data required extensive tidying and cleaning.
That process is documented in the wrangle_report.pdf.

My analysis focused on 
the relationship between how many times a post was retweeted and how often it was favorited; 
the relationship between a dog's age (dog or puppy) and how often the tweet was retweeted;
and 
the relationship between dog breed and how often a tweet was retweeted.
An overview of my findings can be found in act_report.
I allowed myself a little artistic license in this report.
Any opinions expressed therein do not necessarily reflect the opinion's of anyone.

The cleaned and combined data is in 
twiitter_archive_master.csv

The data frame I used for analysis is a paired-down version of the master.csv
It is called dog_analysis.csv